@@ -24,6 +24,7 @@
 		"Other"
 	],
 	"main": "./out/extension.js",
+	"type": "commonjs",
 	"contributes": {
 		"commands": [
 			{
@@ -69,12 +70,15 @@
 		]
 	},
 	"activationEvents": [
-		"*"
+        "onStartupFinished"
 	],
 	"scripts": {
 		"vscode:prepublish": "npm run compile",
+		"start": "npm run compile && code -n --disable-extensions --extensionDevelopmentPath=$PWD",
 		"compile": "tsc -p ./",
 		"watch": "tsc -watch -p ./",
+		"lint": "eslint . --ext .ts",
+		"lint:fix": "eslint . --ext .ts --fix",
 		"pretest": "npm run compile",
 		"test": "node ./out/test/runTest.js"
 	},
@@ -83,11 +87,23 @@
 		"@types/mocha": "^5.2.6",
 		"@types/node": "^10.14.13",
 		"@types/vscode": "^1.36.0",
+		"@typescript-eslint/eslint-plugin": "6.x",
+    "@typescript-eslint/parser": "6.x",
+    "eslint": "8.x",
+    "eslint-config-airbnb-typescript": "17.x",
+    "eslint-config-prettier": "9.x",
+    "eslint-plugin-import": "2.x",
+    "eslint-plugin-node": "11.x",
+    "eslint-plugin-prettier": "5.x",
+    "eslint-plugin-sort-keys-fix": "1.x",
+    "eslint-plugin-typescript-sort-keys": "3.x",
+    "prettier": "3.x",
+		"prettier-eslint": "16.x",
 		"glob": "^7.1.4",
 		"mocha": "^6.1.4",
-		"tslint": "^5.12.1",
-		"typescript": "^3.3.1",
-		"vscode-test": "^1.0.0-next.0"
+		"typescript": "5.3.x",
+		"typescript-eslint": "8.x",
+		"vscode-test": "^1.6.1"
 	},
 	"dependencies": {
 		"async-file": "^2.0.2"
