@@ -1,20 +1,23 @@
-const path = require("path");
-import * as vscode from "vscode";
-import * as fs from "async-file";
-import { configurationParse } from "./configurationParse";
-import { commandsNames } from "../model/commandsNames";
-import defaultJsonFile from "./defaultJsonFile";
-import { filename } from "../model/configurationInfo";
+import * as path from 'path';
+import * as vscode from 'vscode';
+import * as fs from 'async-file';
+import { configurationParse } from './configurationParse';
+import { commandsNames } from '../model/commandsNames';
+import defaultJsonFile from './defaultJsonFile';
+import { filename } from '../model/configurationInfo';
 
 export class ConfigurationFile {
   private configurationFileName = filename;
+
   private globalStoragePath: string;
+
   private fullPath: string;
 
   constructor(private context: vscode.ExtensionContext) {
     this.globalStoragePath = context.globalStoragePath;
-    this.fullPath = this.globalStoragePath + this.configurationFileName;
+    this.fullPath = path.join(this.globalStoragePath, this.configurationFileName);
   }
+
   public async checkIfFileExistAndCrete() {
     if (!(await fs.exists(this.globalStoragePath))) {
       await fs.createDirectory(this.globalStoragePath);
@@ -28,8 +31,8 @@ export class ConfigurationFile {
     await fs.writeTextFile(this.fullPath, defaultJsonFile);
   }
 
-  public watchConfigurationChanges(context: vscode.ExtensionContext) {
-    fs.watchFile(this.fullPath, async (event: any) => {
+  public watchConfigurationChanges() {
+    fs.watchFile(this.fullPath, async () => {
       this.parseFile();
     });
   }
@@ -39,16 +42,10 @@ export class ConfigurationFile {
   }
 
   public editConfigurationFile(context: vscode.ExtensionContext) {
-    let editTerminals = vscode.commands.registerCommand(
-      commandsNames.editConfigurationFile,
-      () => {
-        let terminalsJsonPath = path.join(
-          context.globalStoragePath,
-          this.configurationFileName
-        );
-        vscode.window.showTextDocument(vscode.Uri.file(terminalsJsonPath));
-      }
-    );
+    const editTerminals = vscode.commands.registerCommand(commandsNames.editConfigurationFile, () => {
+      const terminalsJsonPath = path.join(context.globalStoragePath, this.configurationFileName);
+      vscode.window.showTextDocument(vscode.Uri.file(terminalsJsonPath));
+    });
     context.subscriptions.push(editTerminals);
   }
 }
