@@ -1,17 +1,17 @@
-import * as vscode from "vscode";
-import { ConfigurationFile } from "./configurationFile";
+import * as vscode from 'vscode';
+import { ConfigurationFile } from './configurationFile';
 
 export default function activate(context: vscode.ExtensionContext) {
   const configurationFile = new ConfigurationFile(context);
   try {
     configurationFile.checkIfFileExistAndCrete();
     configurationFile.parseFile();
-    configurationFile.watchConfigurationChanges(context);
+    configurationFile.watchConfigurationChanges();
     configurationFile.editConfigurationFile(context);
-  } catch(e){
-      console.log(e);
+  } catch (e) {
+    console.log(e);
     vscode.window.showInformationMessage(
-      "Problem with custom actions, check if you have permissions to write and read files."
+      'Problem with custom actions, check if you have permissions to write and read files.'
     );
   }
 }
