@@ -1,5 +1,4 @@
-export default 
-`[
+export default `[
     {
         "name": "npm start",
         "hide": false,
