@@ -1,12 +1,13 @@
+// eslint-disable-next-line no-shadow
 export enum commandsNames {
+  HasIdAsNameTerminal = 'TerminalStatusBar.HasIdAsNameTerminal',
+  ToggleCurrentTerminal = 'TerminalStatusBar.ToggleCurrentTerminal',
+  ToggleEditStatusBar = 'TerminalStatusBar.ToggleEditStatusBar',
+  ToggleRootTerminal = 'TerminalStatusBar.ToggleRootTerminal',
+  buildInTerminalKill = 'workbench.action.terminal.kill',
   createRootTerminal = 'TerminalStatusBar.createRootTerminal',
+  editConfigurationFile = 'TerminalStatusBar.EditConfigurationFile',
   openInTerminal = 'TerminalStatusBar.OpenInTerminal',
   runInTerminal = 'TerminalStatusBar.RunInTerminal',
-  showTerminal = 'TerminalStatusBar.ShowTerminal',
-  editConfigurationFile = 'TerminalStatusBar.EditConfigurationFile',
-  buildInTerminalKill = 'workbench.action.terminal.kill',
-  ToggleCurrentTerminal = 'TerminalStatusBar.ToggleCurrentTerminal',
-  ToggleRootTerminal = 'TerminalStatusBar.ToggleRootTerminal',
-  ToggleEditStatusBar = 'TerminalStatusBar.ToggleEditStatusBar',
-  HasIdAsNameTerminal = 'TerminalStatusBar.HasIdAsNameTerminal'
+  showTerminal = 'TerminalStatusBar.ShowTerminal'
 }
