@@ -1,10 +1,9 @@
 export interface CustomActionInfo {
+  command: string;
+  hide: boolean;
   name: string;
   path: string;
   when?: string;
-  hide: boolean;
-  command: string;
 }
 
-
-export const filename = "//sideBarActions.json";
+export const filename = 'sideBarActions.json';
