@@ -1,31 +1,34 @@
-import * as vscode from "vscode";
-import { StatusBar } from "./statusBar";
-import { StatusBarsSettings } from "./statusBarsSettings";
+import * as vscode from 'vscode';
+import { StatusBar } from './statusBar';
+import { StatusBarsSettings } from './statusBarsSettings';
 
 export class BuildStatusBar {
   private statusBars: {
     [name: string]: StatusBar;
   } = {};
 
-  constructor(private context: vscode.ExtensionContext, statusBarsSettings: StatusBarsSettings[]) {
+  constructor(
+    private context: vscode.ExtensionContext,
+    statusBarsSettings: StatusBarsSettings[]
+  ) {
     this.Build(statusBarsSettings);
   }
 
   public Build(statusBarsSettings: StatusBarsSettings[]) {
-    statusBarsSettings.forEach( s =>{
+    statusBarsSettings.forEach(s => {
       this.createActionTerminal(s);
       this.registryTheCommands(s);
       this.createStatusBar(s);
     });
   }
 
-  private createActionTerminal(s: StatusBarsSettings){
-      this.statusBars[s.name] = new StatusBar(
-        s.name,
-        this.context,
-        s.priority,
-        s.alignment || vscode.StatusBarAlignment.Left
-      );
+  private createActionTerminal(s: StatusBarsSettings) {
+    this.statusBars[s.name] = new StatusBar(
+      s.name,
+      this.context,
+      s.priority,
+      s.alignment || vscode.StatusBarAlignment.Left
+    );
   }
 
   private registryTheCommands(s: StatusBarsSettings) {
@@ -34,12 +37,10 @@ export class BuildStatusBar {
     });
   }
 
-
-  private createStatusBar(e: StatusBarsSettings){
+  private createStatusBar(e: StatusBarsSettings) {
     this.statusBars[e.name].StatusBar.text = e.text;
     this.statusBars[e.name].StatusBar.tooltip = e.tooltip;
     this.statusBars[e.name].StatusBar.command = e.command;
     this.statusBars[e.name].setVisibility(e.initVisibility);
   }
- 
 }
