@@ -1,38 +1,50 @@
 import * as vscode from 'vscode';
 import { commandsNames } from '../model/commandsNames';
+import { ExTerminal } from '../model/exTerminal';
+
 export class TerminalStatusBarItem {
-  private _statusBarItem: any 
-    = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, -1);
-  private _terminal!: vscode.Terminal;
+  private _statusBarItem: vscode.StatusBarItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, -1);
+  private _terminal!: ExTerminal;
   private _registerCommand: vscode.Disposable;
+  private countClicks: number = 0;
+  private timeoutDisposal: NodeJS.Timeout = null; // eslint-disable-line @typescript-eslint/no-unused-vars
+  private isStillValid = true;
 
-  constructor(terminal: any, reg: vscode.Disposable, hasIdAsName: boolean) {
-    hasIdAsName = true;
+  constructor(terminal: ExTerminal) {
     this._terminal = terminal;
-    this.createStatusBarItem(hasIdAsName);
+    this.createStatusBarItem(true);
     this.addWatchIconWhenTerminalIsBusy(terminal);
-    this._registerCommand = reg;
+    this._registerCommand = vscode.commands.registerCommand(commandsNames.showTerminal + terminal._id, () =>
+      this.clickListeningCommand()
+    );
   }
-
 
   public show(): void {
     this._terminal.show();
   }
-  
+
+  public highlight(): void {
+    // this._terminal.show(); // don't think we need to do this, maybe set color?
+  }
+
+  public isValid(): boolean {
+    return this.isStillValid;
+  }
+
   private createStatusBarItem(hasIdAsName: boolean) {
     if (hasIdAsName) {
-      // @ts-ignore
-      this.setupStatusBarItem(this._terminal.__id__, this._terminal.__id__);
+      this.setupStatusBarItem(this._terminal._id, this._terminal._id);
     } else {
-      // @ts-ignore
-      this.setupStatusBarItem(this._terminal.name, this._terminal.__id__);
+      this.setupStatusBarItem(this._terminal.name, this._terminal._id);
     }
   }
 
-  private setupStatusBarItem(name: string, id: number): void {
+  private setupStatusBarItem(name: string | number, id: number): void {
     this._statusBarItem.text = `$(terminal) ${name}`;
-    this._statusBarItem.tooltip =
-      'Open the terminal \nToggle name by pressing F1 and select "Terminal Status Bar: Toogle appellation of terminal(name or id)"';
+    this._statusBarItem.tooltip = [
+      'Open the terminal ',
+      'Toggle name by pressing F1 and select "Terminal Status Bar: Toggle appellation of terminal(name or id)"'
+    ].join('\n');
     this._statusBarItem.command = commandsNames.showTerminal + id;
     this._statusBarItem.show();
   }
@@ -41,12 +53,13 @@ export class TerminalStatusBarItem {
     Show an watch icon when the terminal do something. 
   */
   private addWatchIconWhenTerminalIsBusy(terminal: vscode.Terminal): void {
-    let delayTimeout: any = null;
+    let delayTimeout: NodeJS.Timeout | null = null;
     return;
-    // @ts-ignore
+    // @ts-expect-error idk
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
     terminal.onDidWriteData(data => {
       if (!this._statusBarItem.text.endsWith('$(watch)')) {
-        this._statusBarItem.text = this._statusBarItem.text + '$(watch)';
+        this._statusBarItem.text = `${this._statusBarItem.text}$(watch)`;
       }
       clearTimeout(delayTimeout);
 
@@ -58,10 +71,25 @@ export class TerminalStatusBarItem {
     });
   }
 
+  private clickListeningCommand() {
+    this.countClicks += 1;
+    if (this.countClicks === 1) {
+      this.show();
+      this.timeoutDisposal = setTimeout(() => {
+        this.countClicks = 0;
+        this.timeoutDisposal = null;
+      }, 650);
+    }
+    if (this.countClicks === 3) {
+      this.timeoutDisposal = null;
+      // three click
+      this.dispose();
+    }
+  }
+
   public changeName(hasIdAsName: boolean) {
     if (hasIdAsName) {
-      // @ts-ignore
-      this._statusBarItem.text = `$(terminal) ${this._terminal.__id__}`;
+      this._statusBarItem.text = `$(terminal) ${this._terminal._id}`;
     } else {
       this._statusBarItem.text = `$(terminal) ${this._terminal.name}`;
     }
@@ -71,5 +99,6 @@ export class TerminalStatusBarItem {
     this._statusBarItem.dispose();
     this._terminal.dispose();
     this._registerCommand.dispose();
+    this.isStillValid = false;
   }
 }
