@@ -6,10 +6,10 @@ import * as vscode from 'vscode';
 // import * as myExtension from '../../extension';
 
 suite('Extension Test Suite', () => {
-	vscode.window.showInformationMessage('Start all tests.');
+  vscode.window.showInformationMessage('Start all tests.');
 
-	test('Sample test', () => {
-		assert.strictEqual([1, 2, 3].indexOf(5), -1);
-		assert.strictEqual([1, 2, 3].indexOf(0), -1);
-	});
+  test('Sample test', () => {
+    assert.strictEqual([1, 2, 3].indexOf(5), -1);
+    assert.strictEqual([1, 2, 3].indexOf(0), -1);
+  });
 });
